taticas prontas para apresentação

+++ gateway/routes/strategies.py
@@ -20,9 +20,11 @@ def create_strategy():
 
         # Junta tática + tempo
         tatics_with_times = [
-            {"name": tatics[i], "time": int(times[i]), "description": description[i]}
+            {"name": tatics[i], "time": float(times[i]), "description": description[i]}
             for i in range(len(tatics))
         ]
+
+        # return f"{tatics_with_times}"
 
         # adicionando o chat_id para cada tática
         for tatics_with_time in tatics_with_times:
